LabReport_github_upload/backend/requirements.txt
fastapi==0.117.0
uvicorn==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
pymysql==1.1.2
passlib[bcrypt]==1.7.4
python-jose==3.5.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.3.0
pdfplumber==0.11.7
python-docx==1.2.0
pytesseract==0.3.13
langchain==0.3.27
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-openai==0.3.30
chromadb==1.0.20
reportlab==4.4.3
